requirements.txt
django>=4.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pandas
openpyxl